Breaking stuff into separate files, refactoring, changing block to pass as ref, and to "clear" a blockchain from commandline
